+++ KeyboardHookState.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace uclliu
 {
     public static class KeyboardHookMessage
@@ -6,6 +8,7 @@ namespace uclliu
         public const int WmKeyUp = 0x0101;
         public const int WmSysKeyDown = 0x0104;
         public const int WmSysKeyUp = 0x0105;
+        public const int LowLevelInjectedFlag = 0x10;
 
         public static bool IsKeyDown(int message)
         {
@@ -15,6 +18,16 @@ namespace uclliu
         public static bool IsKeyUp(int message)
         {
             return message == WmKeyUp || message == WmSysKeyUp;
+        }
+
+        public static bool IsInjected(int flags)
+        {
+            return (flags & LowLevelInjectedFlag) == LowLevelInjectedFlag;
+        }
+
+        public static bool IsInjectedByUcl(int flags, IntPtr extraInfo)
+        {
+            return IsInjected(flags) && extraInfo == UnicodeSendInputOutput.UclExtraInfo;
         }
     }